requirements.txt
#
# Requirements for wger for production
#
Django==1.9
django-recaptcha==1.3.1
reportlab==3.3.0
django-mobile==0.7.0
django-formtools>=1.0,<1.1
bleach==3.1.0
python-mimeparse==1.6.0
Pillow==5.4.1
easy-thumbnails==2.6
django-compressor==2.2
icalendar==4.0.3
django-bootstrap-breadcrumbs==0.9.1
Sphinx==1.8.5
django-sortedm2m==1.5.0
django-bower==5.2.0
invoke==1.2.0
requests==2.21.0
whitenoise==4.1.2
gunicorn==19.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
social-auth-app-django==3.1.0
social-auth-core==3.2.0
# REST API
djangorestframework>=3.2,<3.3
django-filter==1.1.0
django-tastypie==0.13.3
django-cors-headers==2.4.1
django-tabulate==0.1.1
# Python3 compatibility
six==1.12.0